==> analysis/6-BMDM.Rmd <==
---
title: "Zoom in on BMDM"
date: University of Bern, DBMR, Visceral and Transplantation Surgery lab
author: Tess Brodie
output: 
  html_document:
    toc_depth: 2
---


```{r setup, include = FALSE, dev='png'}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)
knitr::opts_chunk$set(dev = 'png')
```

## Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(CATALYST)
library(dplyr)
library(workflowr)
library(flowCore)
library(ggplot2)
library(diffcyt)
library(cowplot)
library(viridis)
library(RColorBrewer)
```

### Load data

```{r load-data}
sce <- readRDS(file.path("output", "sce.rds"))
```

# Filter on bone marrow derived macrophages (BMDM)

```{r Filter_BMDM}
sceB <- filterSCE(sce, k = "mm",
               cluster_id %in% "BMDM")

plot_grid(
  plotDR(sce, dr= "TSNE", color_by = "mm"),
  plotDR(sceB, dr= "TSNE", color_by = "mm"))
```

# Target expression over time

```{r ExprHeatmap_BMDM}
plotExprHeatmap(sceB,
    row_anno = FALSE,   # don't annotate samples
    row_clust = FALSE,  # keep samples in original order
    col_clust = TRUE,  # keep markers in original order
    bin_anno = FALSE,   # don't annotate bins
    bars = FALSE,       # don't include sample sizes
    scale = "last")     # aggregate, then scale
```

# Number per sample

```{r BMDM_count}
plotCounts(sceB, 
           group_by = "sample_id", 
           color_by = "condition")
```

# tSNE by condition

```{r TSNE_condition}
sceB <- runDR(sceB, dr = "TSNE", features = NULL, cells = 1000)
plotDR(sceB, dr = "TSNE", color_by = "condition")
```

## Faceted

```{r TSNE_condition_facet}
plotDR(sceB, dr = "TSNE", color_by = "condition", facet_by = "condition")
```

# Heatmap Type markers

```{r plotDR-expression_Type, fig.width = 12, fig.height = 5}
p <- plotDR(sceB, dr = "TSNE", color_by = type_markers(sce))
p$facet$params$ncol <- 4; p
```

## ...split by condition {.tabset}

```{r plotDR-expression-condition, results = "asis", fig.width = 12, fig.height = 5}
for (i in state_markers(sce)) {
    p <- plotDR(sceB, dr = "TSNE", color_by = i, facet_by = "condition")
    cat("### ", i, "\n"); print(p); cat("\n\n")
}
```


# Clustered into 10 subsets
```{r cluster_10, message = FALSE}
sceB <- cluster(sceB, maxK = 10)
```

# Expression heatmap_10
```{r plotExprHeatmap_10, fig.width = 6, fig.height = 5}
plotExprHeatmap(sceB, features = type_markers(sce), 
    by = "cluster_id", k = "meta10", row_anno = FALSE, bars = TRUE)
```

# Clusters overlaid on TSNE_10
```{r TSNE_meta10}
plotDR(sceB, dr = "TSNE", color_by = "meta10")
```

## Faceted

```{r TSNE_cluster_facet}
plotDR(sceB, dr = "TSNE", color_by = "meta10", facet_by = "condition")
```

# Frequency by cluster_10

```{r Freqheatmap_cluster10}
plotFreqHeatmap(sceB, 
    k = "meta10",
    hm_pal = rev(hcl.colors(10, "RdBu")),
    k_pal = hcl.colors(7, "Zissou 1"),
    col_clust = FALSE, col_anno = FALSE,
    row_anno = FALSE,
    bars = TRUE, perc = TRUE)
```

# Abundance stack plots

```{r Abundance stack}
plotAbundances(sceB, k = "meta10", by = "sample_id", group_by = "condition")
```

# Expression of KC markers in BMDM subsets
To address if some BMDM subsets may be transforming into KCs

```{r multiheatmap_kCmarkersBMDM}
plotMultiHeatmap(sceB, 
    k = "meta10", scale = "never",
    hm1 = FALSE, hm2 = c("F4_80", "CD11b", "CD206"),
    row_anno = FALSE, col_clust = FALSE,
    hm2_pal = c("grey95", "black"))
```

# DS analysis
## Differential state significance testing
```{r DS}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DS analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
  # create contrast matrix
  contrast <- createContrast(as.numeric(colnames(design) == i))
    
  # run DS analysis & extract results table
  res_DS <- diffcyt(sceB, clustering_to_use = "meta10",
                  analysis_type = "DS", method_DS = "diffcyt-DS-limma",
                  design = design, contrast = contrast, verbose = FALSE)

  # extract results table
  tbl_DS <- rowData(res_DS$res)

  # write table to .csv
    fn <- file.path("output", sprintf("DS-results_KCs_%s.csv", i))
    write.csv(tbl_DS, fn, row.names = FALSE)
    
  # subset control and timepoint of interest
    sub <- filterSCE(sceB, condition %in% c("control", !!i))
    
  # plot differential heatmap
    p <- plotDiffHeatmap(sub, tbl_DS, top_n = 30, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
} 
```

## Differential Abundances significance testing

# DA analysis {.tabset}
```{r da, results = "asis", fig.width = 8, fig.height = 6}
# create design matrix
design <- createDesignMatrix(ei(sceB), "condition")
colnames(design) <- gsub("condition", "", colnames(design))

# run DA analysis for control vs. each timepoint
for (i in colnames(design)[-1]) {
    # create contrast matrix
    contrast <- createContrast(as.numeric(colnames(design) == i))
    
    # run DA analysis
    res <- diffcyt(sceB, clustering_to_use = "meta10",
        analysis_type = "DA", method_DA = "diffcyt-DA-edgeR",
        design = design, contrast = contrast, verbose = FALSE)
    
    # extract results table
    tbl <- rowData(res$res)
    
    # write table to .csv
    fn <- file.path("output", sprintf("DA-results_KCs_%s.csv", i))
    write.csv(tbl, fn, row.names = FALSE)
    
    # subset control and timepoint of interest
    sub <- filterSCE(sceB, condition %in% c("control", !!i))
    
    # plot differential heatmap
    p <- plotDiffHeatmap(sub, tbl, all = TRUE, col_anno = "condition")
    cat("## ", i, "\n"); print(p); cat("\n\n")
}
```
